+++ h2gis-functions/src/main/java/org/h2gis/functions/factory/H2GISFunctions.java
@@ -54,6 +54,9 @@ import org.h2gis.functions.spatial.aggregate.ST_Collect;
 import org.h2gis.functions.spatial.aggregate.ST_LineMerge;
 import org.h2gis.functions.spatial.buffer.*;
 import org.h2gis.functions.spatial.clean.ST_MakeValid;
+import org.h2gis.functions.spatial.clusters.ST_ClusterDBScan;
+import org.h2gis.functions.spatial.clusters.ST_ClusterIntersecting;
+import org.h2gis.functions.spatial.clusters.ST_ClusterWithin;
 import org.h2gis.functions.spatial.convert.*;
 import org.h2gis.functions.spatial.coverage.ST_CoverageUnion;
 import org.h2gis.functions.spatial.create.*;
@@ -351,7 +354,10 @@ public class H2GISFunctions {
                 new ST_M(),
                 new ST_TriangulatePolygon(),
                 new ST_MaximumInscribedCircleRadius(),
-                new ST_SimplifyPolygonHull()
+                new ST_SimplifyPolygonHull(),
+                new ST_ClusterDBScan(),
+                new ST_ClusterIntersecting(),
+                new ST_ClusterWithin()
         };
     }
 
@@ -501,7 +507,8 @@ public class H2GISFunctions {
                     st.execute("DROP AGGREGATE IF EXISTS " + functionAlias);
                 }
                 st.execute("CREATE FORCE AGGREGATE IF NOT EXISTS " + functionAlias + " FOR \"" + packagePrepend + functionClass + "\"");
-        } else {
+        }
+        else {
                 throw new SQLException("Unsupported function "+functionClass);
         }
     }